requirements.txt
sentence-transformers
faiss-cpu
numpy
pandas
requests
tqdm
transformers
torch
streamlit
